requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
contourpy==1.0.5
cryptography==41.0.3
cycler==0.11.0
DateTime==5.3
Flask==3.0.0
fonttools==4.25.0
frozenlist==1.4.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.61.0
greenlet==3.0.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lambda-packager==1.2.2
lxml==4.9.3
MarkupSafe==2.1.1
matplotlib==3.8.0
MouseInfo==0.1.3
mpld3==0.5.9
multidict==6.0.2
munkres==1.1.4
numexpr==2.8.4
numpy==1.26.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.1
pip==23.2.1
playwright==1.39.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pyee==11.0.1
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3.post1
pyu2f==0.1.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
rubicon-objc==0.4.7
selenium==3.141.0
setuptools==68.0.0
six==1.16.0
soupsieve==2.4
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==3.0.1
wheel==0.41.2
yarl==1.8.1
zope.interface==6.1
